pkg/bgp: add non-parsing variants for all types
All of the BGP types used internally had New* functions that took a
string as a parameter to the parse it. It's fine and all unless it's in
a hot loop doing NewThing(address.String()), costing two allocations
(string and net.IP once parsed), a serialization and a parse.

This commit basically hunted down those functions, turned them into a
Parse* variant, and introduced New* functions that accept a net.IP. And
callsites were examined to avoid such multi-allocations everywhere, and
string round-trips.