table: make tableKey faster and use less memory for EVPN
When using EVPN, the table key is the string representation of the
prefix. With a large number of routes, this quickly adds up (observed to
undreds of megabytes with 500k evpn type-2 routes).

This also has the benefit of making hashing, accesses, updates, etc
slightly faster by using smaller keys, and not requiring string
formatting. With a large number of routes, it adds up a bit.